+++ includes/class-plugin-name-i18n.php
@@ -1,5 +1,5 @@
 <?php
-namespace PluginName;
+namespace Plugin_Name;
 /**
  * Define the internationalization functionality
  *