+++ includes/class-plugin-name-loader.php
@@ -1,5 +1,5 @@
 <?php
-namespace PluginName;
+namespace Plugin_Name;
 /**
  * Register all actions and filters for the plugin
  *
@@ -95,8 +95,8 @@ class Plugin_Name_Loader {
    * Add a new shortcode to the collection to be registered with WordPress.
    *
    * @since    2.0.0
-   * @param    string               $hook             The name of the WordPress filter that is being registered.
-   * @param    object               $component        A reference to the instance of the object on which the filter is defined.
+   * @param    string               $hook             The name of the WordPress shortcode that is being registered.
+   * @param    object               $component        A reference to the instance of the object on which the shortcode is defined.
    * @param    string               $callback         The name of the function definition on the $component.
    * @param    int                  $priority         Optional. The priority at which the function should be fired. Default is 10.
    * @param    int                  $accepted_args    Optional. The number of arguments that should be passed to the $callback. Default is 1